backend/requirements.txt
django
django-enumfields
djangorestframework
psycopg2==2.9.3